% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block2array.R
\name{block2array}
\alias{block2array}
\title{Convert block matrix to array}
\usage{
block2array(M, N, TT)
}
\arguments{
\item{M}{NTxNT matrix with NxN sub-blocks}

\item{N}{dim of sub-blocks}

\item{TT}{Number of sub-blocks in each row (column)}
}
\value{
an array of dim [N,T,N,T] where [, i, , j] gives the (i,j)th NxN sub-block
}
\description{
Convert NTxNT matrix with NxN sub-blocks into [N,T,N,T] array
}
