% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builD.R
\name{builD}
\alias{builD}
\title{Put together inverse covariance matrix of overdifferenced signals.
    notated as $D$ in (Livsey and McElroy, 2018+ - EMsigex)}
\usage{
builD(mdl)
}
\arguments{
\item{mdl}{sigex mdl object}
}
\value{
List of inverse covariance matrices
}
\description{
Put together inverse covariance matrix of overdifferenced signals.
    notated as $D$ in (Livsey and McElroy, 2018+ - EMsigex)
}
