% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demean.R
\name{demean}
\alias{demean}
\title{Demean vector time series}
\usage{
demean(x)
}
\arguments{
\item{x}{matrix vector valued time series (assumes dim < sample size)}
}
\value{
same dim matrix as supplied with sample mean removed
}
\description{
takes a sample of size (T) x dim (N) matrix (or its transpose) and removes the
sample means of each column (row).
}
