% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMcritical_1-B12.R
\name{EMcritical_1_B12}
\alias{EMcritical_1_B12}
\title{Calculate updated Sigma values at EM critical point}
\usage{
EMcritical_1_B12(j, Sig, lMS)
}
\arguments{
\item{j}{component of mdl}

\item{lMS}{list with lMS[[1]] = M an [N,N,T,T] array and lMS[[2]] = S[T,N]
matrix}
}
\value{
NxN updated Sigma matrix
}
\description{
Calculate updated Sigma values at EM critical point
}
