% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMsigex.R
\name{EMsigex}
\alias{EMsigex}
\title{Perform multivariate signal extraction with *sigex* via the EM algoritm}
\usage{
EMsigex(data, mdl, transform = "none")
}
\arguments{
\item{data}{multivariate time series as a matrix (sampleSize x dim)}

\item{mdl}{sigex model object}

\item{transform}{transform for data. Default "none"}
}
\value{
list of parameter estimates
}
\description{
Perform multivariate signal extraction with *sigex* via the EM algoritm
}
