% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate Data from Trend + Seasonal + Irregular model
Assumes (1 - B) reduces trend to stationary and (1-B)^{12} reduces
the seasonal component to stationary}
\usage{
gen_data(N, TT)
}
\arguments{
\item{N}{dimensionality}

\item{TT}{sample size}
}
\value{
multivariate time series of sigex form
}
\description{
Generate Data from Trend + Seasonal + Irregular model
Assumes (1 - B) reduces trend to stationary and (1-B)^{12} reduces
the seasonal component to stationary
}
