% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_seasComp.R
\name{gen_seasComp}
\alias{gen_seasComp}
\title{Generate multivariate seasonal component}
\usage{
gen_seasComp(n, Phi, Sig, burn = 10^3, seasonal.period = 12)
}
\arguments{
\item{n}{length of series}

\item{Phi}{seasonal autoregressive parameter}

\item{Sig}{covariance martrix white noise component}

\item{burn}{burn in (defaults to 1000)}

\item{seasonl.period}{periodicity of component (defaults to 12)}
}
\value{
Ndim x n matrix of observations
}
\description{
generates a component that is seasonally (12) difference stationary
}
