% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param2cor.R
\name{param2cor}
\alias{param2cor}
\title{convert param list to correlation matrix list}
\usage{
param2cor(param)
}
\arguments{
\item{param}{parameter list in sigex form}
}
\value{
list of NxN correlation matricies
}
\description{
convert param list to correlation matrix list
}
