% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param2sig.R
\name{param2sig}
\alias{param2sig}
\title{convert param list to Sigma matrix list}
\usage{
param2sig(param)
}
\arguments{
\item{param}{parameter list in sigex form}
}
\value{
list of NxN Sigma matricies
}
\description{
convert param list to Sigma matrix list
}
