% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row2sig.R
\name{row2sig}
\alias{row2sig}
\title{convert vector of stored values to Sig list form}
\usage{
row2sig(v, N)
}
\arguments{
\item{v}{vector of unlisted values}

\item{N}{dimension of output matricies}
}
\value{
list of Sig matricies
}
\description{
convert vector of stored values to Sig list form
}
