% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2param.R
\name{sig2param}
\alias{sig2param}
\title{Convert list of Sigma matricies to sigex param list}
\usage{
sig2param(Sig)
}
\arguments{
\item{Sig}{list with each element of list being an NxN sigma matrix}
}
\value{
updated param list
}
\description{
Convert list of Sigma matricies to sigex param list
}
