% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigexRun.R
\name{sigexRun}
\alias{sigexRun}
\title{Runs sigex for EM algorithm}
\usage{
sigexRun(param)
}
\arguments{
\item{param}{parameter list is sigex form}
}
\value{
list with M (error matricies) and S (signal estimates)
}
\description{
Runs sigex for EM algorithm
}
