% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigexRun_1-B12.R
\name{sigexRun_1_B12}
\alias{sigexRun_1_B12}
\title{Runs sigex for EM algorithm}
\usage{
sigexRun_1_B12(param, data, mdl)
}
\arguments{
\item{param}{parameter list in sigex form}

\item{data}{mts object (sample size) x (dim)}

\item{mdl}{mdl object in sigex form}
}
\value{
list with M (error matricies) and S (signal estimates)
}
\description{
Runs sigex for EM algorithm
}
