% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Calculate trace of a square matrix}
\usage{
tr(m)
}
\arguments{
\item{m}{square matrix to calculate trace of}
}
\value{
trace of given matrix
}
\description{
Calculate trace of a square matrix
}
