% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMiterate.R
\name{EMiterate}
\alias{EMiterate}
\title{single iteration of EM algorithm}
\usage{
EMiterate(Sig, lMS, data, mdl)
}
\arguments{
\item{Sig}{list with each element of list being an NxN sigma matrix}

\item{lMS}{list with lMS[[1]] = M an [N,N,T,T] array and lMS[[2]] = S[T,N]
matrix}
}
\value{
updated values of all inputs
}
\description{
single iteration of EM algorithm
}
