% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2lik.R
\name{sig2lik}
\alias{sig2lik}
\title{output likelihood value for given Sigma values}
\usage{
sig2lik(Sig)
}
\arguments{
\item{Sig}{List of Sigma's to output likelihood of}
}
\value{
likelihood value
}
\description{
output likelihood value for given Sigma values
}
