% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigexRun.R, R/sigexRun_1-B12.R
\name{sigexRun}
\alias{sigexRun}
\alias{sigexRun}
\title{Runs sigex for EM algorithm}
\usage{
sigexRun(param)

sigexRun(param)
}
\arguments{
\item{param}{parameter list in sigex form}

\item{data}{data matrix (dim x sample_size)}

\item{mdl}{sigex mdl}

\item{param}{parameter list in sigex form}
}
\value{
list with M (error matricies) and S (signal estimates)

list with M (error matricies) and S (signal estimates)
}
\description{
Runs sigex for EM algorithm

Runs sigex for EM algorithm
}
